+++ anitection/lib/screens/institution_detail/institution_screen.dart
@@ -1,4 +1,5 @@
 import 'package:anitection/components/nav_up_button.dart';
+import 'package:anitection/components/stroke_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -327,45 +328,3 @@ class _Text extends StatelessWidget {
   }
 }
 
-class StrokeText extends StatelessWidget {
-  final String text;
-  final double strokeWidth;
-  final Color textColor;
-  final Color strokeColor;
-  final TextStyle? textStyle;
-  final List<Shadow>? shadows;
-
-  const StrokeText(
-      {Key? key,
-      required this.text,
-      this.strokeWidth = 1,
-      this.strokeColor = Colors.black,
-      this.textColor = Colors.white,
-      this.textStyle,
-      this.shadows})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        Text(
-          text,
-          style: TextStyle(
-            foreground: Paint()
-              ..style = PaintingStyle.stroke
-              ..strokeWidth = strokeWidth
-              ..color = strokeColor,
-            shadows: shadows,
-          ).merge(textStyle),
-        ),
-        Text(
-          text,
-          style: TextStyle(
-            color: textColor,
-          ).merge(textStyle),
-        ),
-      ],
-    );
-  }
-}
